EmailSummary/requirements.txt
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
filelock==3.12.2
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
numpy==1.25.0
packaging==23.1
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.6.3
urllib3==2.0.3
